+++ py/desitarget/mtl.py
@@ -38,8 +38,8 @@ def make_mtl(targets, obscon, zcat=None, trim=False, scnd=None):
         any more observations.  If ``False``, include every input target.
     scnd : :class:`~numpy.array`, `~astropy.table.Table`, optional
         A set of secondary targets associated with the `targets`. As with
-        the `target` must include at least ``TARGETID``, ``DESI_TARGET``,
-        ``NUMOBS_INIT``, ``PRIORITY_INIT`` or the corresponding SV columns.
+        the `target` must include at least ``TARGETID``, ``NUMOBS_INIT``,
+        ``PRIORITY_INIT`` or the corresponding SV columns.
         The secondary targets will be padded to have the same columns
         as the targets, and concatenated with them.
 
@@ -67,6 +67,9 @@ def make_mtl(targets, obscon, zcat=None, trim=False, scnd=None):
         for col in sharedcols:
             padit[col] = scnd[col]
         targets = np.concatenate([targets, padit])
+        # APC Propagate a flag on which targets came from scnd
+        is_scnd = np.repeat(False, len(targets))
+        is_scnd[-nrows:] = True
         log.info('Done with padding...t={:.1f}s'.format(time()-start))
 
     # ADM determine whether the input targets are main survey, cmx or SV.
@@ -160,6 +163,11 @@ def make_mtl(targets, obscon, zcat=None, trim=False, scnd=None):
 
     # - Set the OBSCONDITIONS mask for each target bit.
     obsconmask = set_obsconditions(targets)
+
+    # APC obsconmask will now be incorrect for secondary-only targets. Fix this
+    # APC using the mask on secondary targets.
+    if scnd is not None:
+        obsconmask[is_scnd] = set_obsconditions(targets[is_scnd], scnd=True)
 
     # ADM set up the output mtl table.
     mtl = Table(targets)